+++ src/test/java/com/braintreegateway/integrationtest/CreditCardIT.java
@@ -882,7 +882,6 @@ public class CreditCardIT extends IntegrationTest implements MerchantAccountTest
         RiskData riskData = verification.getRiskData();
         assertNotNull(riskData);
 
-        assertEquals("Approve", riskData.getDecision());
         assertFalse(riskData.getDeviceDataCaptured());
         assertNotNull(riskData.getId());
         assertNotNull(riskData.getFraudServiceProvider());