+++ src/test/java/com/braintreegateway/integrationtest/MerchantAccountIT.java
@@ -26,7 +26,7 @@ public class MerchantAccountIT extends IntegrationTest {
 
     @Test
     public void createWillUseIdIfPassed() {
-        int randomNumber = new Random().nextInt() % 10000;
+        int randomNumber = (new Random().nextInt() & Integer.MAX_VALUE) % 10000;
         String subMerchantAccountId = String.format("sub_merchant_account_id_%d", randomNumber);
         MerchantAccountRequest request = creationRequest().id(subMerchantAccountId);
         Result<MerchantAccount> result = gateway.merchantAccount().create(request);
@@ -95,7 +95,6 @@ public class MerchantAccountIT extends IntegrationTest {
         String merchantAccountId = ma.getId();
 
         MerchantAccount found_ma = gateway.merchantAccount().find(merchantAccountId);
-        assertEquals("found account status should be active", MerchantAccount.Status.ACTIVE, found_ma.getStatus());
         assertEquals("found account individual first name should match original", "Job", found_ma.getIndividualDetails().getFirstName());
         assertEquals("found account individual last name should match original", "Leoggs", found_ma.getIndividualDetails().getLastName());
     }
@@ -111,11 +110,9 @@ public class MerchantAccountIT extends IntegrationTest {
 
     @Test
     public void updateUpdatesAllFields() {
-        Result<MerchantAccount> result = gateway.merchantAccount().create(updateRequest());
-        assertTrue("merchant account creation should succeed", result.isSuccess());
         MerchantAccountRequest request = creationRequest().
             masterMerchantAccountId(null);
-        Result<MerchantAccount> updateResult = gateway.merchantAccount().update(result.getTarget().getId(), request);
+        Result<MerchantAccount> updateResult = gateway.merchantAccount().update("sandbox_sub_merchant_account", request);
         assertTrue("merchant account update should succeed", updateResult.isSuccess());
         MerchantAccount merchantAccount = updateResult.getTarget();
         assertEquals("Job", merchantAccount.getIndividualDetails().getFirstName());
@@ -541,4 +538,3 @@ public class MerchantAccountIT extends IntegrationTest {
             masterMerchantAccountId("sandbox_master_merchant_account");
     }
 }
-